Import first version of the script-extractor
This program parses the xml files generated by the poppler tool pdftohtml. It
is already able the guess the ident-levels of stage directions, speakers and
dialog. It can also group paragraphs together by looking at the line spacing.

This version does not yet split the script into distinct scenes or provide the
script in the planned xml format. Instead it prints all lines annotated to
stdout for further refinement.