Generate a matrix of appearances from a script
The matrix counts the number of times a character appears together with
another in a script.  The matrix is intended to be consumed by d3's
chord layout.

This only works with preprocessed scripts.  (I.e. the `toy-story-3.txt`
file was generated by running the extractor and than adding "scene
markers" manually.  Scenes are delimited by lines beginning with
`-- scene:`.)