Implement proper module interfaces and add some basic documentation
Previously several internal functions of the parse module had to be public.
This commit adds a new public wrapper function that takes a reader and returns
the parsed Script without exposing internal details.

This commit also restructures the modules so that the public interfaces live
at the top of the module files.